make stable hashtable expansion
Regenerate slots basing on the array of stored nodes.

Sorting makes the top node of a linked list down in a new slot. It leads
to random node order after expansions, especially for the same hash but
different nodes. Here, for example: 2+ rules for a same domain as nodes,
any one could be on the top.